Fix poethepoet usage in GitHub Actions
poethepoet is a task runner and a plugin for poetry to manage dependencies.
By installing it as a plugin, you can write `poetry run poe xxx` instead of `poetry poe xxx`.
I used it as a plugin in GitHub Actions, but it doesn't work well with cache, so I'll write `poetry run poe xxx` instead.