Try nbmake
nbmake is a plugin for pytest to test Jupyter Notebook input and output.

 ## Motivation

Jupyter Notebook is an essential item in the field of machine learning.
On the other hand, the raw Jupyter Notebook is extremely incompatible with tests.
With nbmake, you can test Jupyter Notebook input and output as a pytest plugin.