lcm_cat
`lcm_cat` is an operation that concatenates multiple tensors.
Normally, `torch.cat` requires that the number of dimensions of the tensors to be concatenated is the same.
lcm_cat can concatenate tensors with different numbers of dimensions.
The least common multiple of the dimensions of the tensors to be concatenated is calculated, and the tensors are expanded to match it before concatenation.

```python
    asis = lcm_cat(
        [
            torch.full((2, 3, 2, 3), 8),
            torch.full((1, 3, 3, 1), 9),
        ]
    )
    tobe = torch.tensor(
        [
            [
                [8, 8, 8, 8, 8, 8, 9, 9, 9],
                [8, 8, 8, 8, 8, 8, 9, 9, 9],
                [8, 8, 8, 8, 8, 8, 9, 9, 9],
            ],
            [
                [8, 8, 8, 8, 8, 8, 9, 9, 9],
                [8, 8, 8, 8, 8, 8, 9, 9, 9],
                [8, 8, 8, 8, 8, 8, 9, 9, 9],
            ],
        ]
    )
    assert torch.all(asis == tobe)
    assert asis.shape == tobe.shape == (2, 3, 9)
```