Workaround for poetry and debugpy compatibility problem
A compatibility problem with poetry v1.4.1 and debugpy has occurred.
In my opinion, the problem occurs because the hash value locked by poetry is different from the hash value locked by debugpy when verdoring pydevd in debugpy.
There are no other problems with many other vendoring packages, so it seems that the problem is with debugpy?
However, it doesn't seem to be resolved soon, so for now, change the poetry configuration as a workaround.