Add example for pick_max_steps algorithm
Pick the maximum item from given items and decrease it by 1.
It will repeat until all items are less than or equal to 0.
This means every item will be picked at least once.
The result is the sequence of the keys of the items.

Example:
When the items are `A: 4.2, B: 3.5, C: 2.8, D: 1.1`
The result will be `A, B, A, C, B, A, C, B, A, D, C, B, A, D`
```
0: A: 4.2, B: 3.5, C: 2.8, D: 1.1 ... pick A
1: A: 3.2, B: 3.5, C: 2.8, D: 1.1 ... pick B
1: A: 3.2, B: 2.5, C: 2.8, D: 1.1 ... pick A
2: A: 3.2, B: 2.5, C: 1.8, D: 1.1 ... pick C
3: A: 2.2, B: 2.5, C: 1.8, D: 1.1 ... pick B
4: A: 2.2, B: 1.5, C: 1.8, D: 1.1 ... pick A
5: A: 1.2, B: 1.5, C: 1.8, D: 1.1 ... pick C
5: A: 1.2, B: 1.5, C: 0.8, D: 1.1 ... pick B
6: A: 0.2, B: 1.5, C: 0.8, D: 1.1 ... pick A
7: A: 0.2, B: 0.5, C: 0.8, D: 1.1 ... pick D
8: A: 0.2, B: 0.5, C: 0.8, D: 0.1 ... pick C
9: A: 0.2, B: 0.5, C: -0.2, D: 0.1 ... pick B
10: A: 0.2, B: -0.5, C: -0.2, D: 0.1 ... pick A
11: A: -0.8, B: -0.5, C: -0.2, D: 0.1 ... pick D
12: A: -0.8, B: -0.5, C: -0.2, D: -0.9 ... End
```